Move sub-classes of "UpDownBase", "WebBrowser" and "WebBrowserBase" classes to their own files
