+++ src/System.Windows.Forms/src/System/Windows/Forms/WebBrowser.cs
@@ -12,7 +12,6 @@ using System.Runtime.InteropServices;
 using System.Text;
 using static Interop;
 using static Interop.Mshtml;
-using IComDataObject = System.Runtime.InteropServices.ComTypes.IDataObject;
 
 namespace System.Windows.Forms
 {
@@ -24,7 +23,7 @@ namespace System.Windows.Forms
     [Docking(DockingBehavior.AutoDock)]
     [SRDescription(nameof(SR.DescriptionWebBrowser))]
     [Designer("System.Windows.Forms.Design.WebBrowserDesigner, " + AssemblyRef.SystemDesign)]
-    public class WebBrowser : WebBrowserBase
+    public partial class WebBrowser : WebBrowserBase
     {
         // Reference to the native ActiveX control's IWebBrowser2
         // Do not reference this directly. Use the AxIWebBrowser2
@@ -1433,410 +1432,6 @@ namespace System.Windows.Forms
                     throw new InvalidOperationException(SR.WebBrowserNoCastToIWebBrowser2);
                 }
                 return axIWebBrowser2;
-            }
-        }
-
-        //
-        // WebBrowserSite class:
-        //
-        /// <summary>
-        ///  Provides a default WebBrowserSite implementation for use in the CreateWebBrowserSite
-        ///  method in the WebBrowser class.
-        /// </summary>
-        protected class WebBrowserSite : WebBrowserSiteBase, IDocHostUIHandler
-        {
-            /// <summary>
-        ///  Creates an instance of the <see cref='WebBrowserSite'/> class.
-        /// </summary>
-            public WebBrowserSite(WebBrowser host) : base(host)
-            {
-            }
-
-            // IDocHostUIHandler Implementation
-            unsafe HRESULT IDocHostUIHandler.ShowContextMenu(uint dwID, Point* pt, object pcmdtReserved, object pdispReserved)
-            {
-                WebBrowser wb = (WebBrowser)Host;
-
-                if (wb.IsWebBrowserContextMenuEnabled)
-                {
-                    // let MSHTML display its UI
-                    return HRESULT.S_FALSE;
-                }
-
-                if (pt is null)
-                {
-                    return HRESULT.E_INVALIDARG;
-                }
-
-                if (pt->X == 0 && pt->Y == 0)
-                {
-                    // IDocHostUIHandler::ShowContextMenu sends (0,0) when the context menu is invoked via the keyboard
-                    // make it (-1, -1) for the WebBrowser::ShowContextMenu method
-                    pt->X = -1;
-                    pt->Y = -1;
-                }
-                wb.ShowContextMenu(pt->X, pt->Y);
-                // MSHTML should not display its context menu because we displayed ours
-                return HRESULT.S_OK;
-            }
-
-            unsafe HRESULT IDocHostUIHandler.GetHostInfo(DOCHOSTUIINFO* pInfo)
-            {
-                if (pInfo is null)
-                {
-                    return HRESULT.E_POINTER;
-                }
-
-                WebBrowser wb = (WebBrowser)Host;
-
-                pInfo->dwDoubleClick = DOCHOSTUIDBLCLK.DEFAULT;
-                pInfo->dwFlags = DOCHOSTUIFLAG.NO3DOUTERBORDER |
-                               DOCHOSTUIFLAG.DISABLE_SCRIPT_INACTIVE;
-
-                if (wb.ScrollBarsEnabled)
-                {
-                    pInfo->dwFlags |= DOCHOSTUIFLAG.FLAT_SCROLLBAR;
-                }
-                else
-                {
-                    pInfo->dwFlags |= DOCHOSTUIFLAG.SCROLL_NO;
-                }
-
-                if (Application.RenderWithVisualStyles)
-                {
-                    pInfo->dwFlags |= DOCHOSTUIFLAG.THEME;
-                }
-                else
-                {
-                    pInfo->dwFlags |= DOCHOSTUIFLAG.NOTHEME;
-                }
-
-                return HRESULT.S_OK;
-            }
-
-            HRESULT IDocHostUIHandler.EnableModeless(BOOL fEnable)
-            {
-                return HRESULT.E_NOTIMPL;
-            }
-
-            HRESULT IDocHostUIHandler.ShowUI(
-                uint dwID,
-                Ole32.IOleInPlaceActiveObject activeObject,
-                Ole32.IOleCommandTarget commandTarget,
-                Ole32.IOleInPlaceFrame frame,
-                Ole32.IOleInPlaceUIWindow doc)
-            {
-                return HRESULT.S_FALSE;
-            }
-
-            HRESULT IDocHostUIHandler.HideUI()
-            {
-                return HRESULT.E_NOTIMPL;
-            }
-
-            HRESULT IDocHostUIHandler.UpdateUI()
-            {
-                return HRESULT.E_NOTIMPL;
-            }
-
-            HRESULT IDocHostUIHandler.OnDocWindowActivate(BOOL fActivate)
-            {
-                return HRESULT.E_NOTIMPL;
-            }
-
-            HRESULT IDocHostUIHandler.OnFrameWindowActivate(BOOL fActivate)
-            {
-                return HRESULT.E_NOTIMPL;
-            }
-
-            unsafe HRESULT IDocHostUIHandler.ResizeBorder(RECT* rect, Ole32.IOleInPlaceUIWindow doc, BOOL fFrameWindow)
-            {
-                return HRESULT.E_NOTIMPL;
-            }
-
-            HRESULT IDocHostUIHandler.GetOptionKeyPath(string[] pbstrKey, uint dw)
-            {
-                return HRESULT.E_NOTIMPL;
-            }
-
-            HRESULT IDocHostUIHandler.GetDropTarget(Ole32.IDropTarget pDropTarget, out Ole32.IDropTarget ppDropTarget)
-            {
-                // Set to null no matter what we return, to prevent the marshaller
-                // from having issues if the pointer points to random stuff.
-                ppDropTarget = null;
-                return HRESULT.E_NOTIMPL;
-            }
-
-            HRESULT IDocHostUIHandler.GetExternal(out object ppDispatch)
-            {
-                WebBrowser wb = (WebBrowser)Host;
-                ppDispatch = wb.ObjectForScripting;
-                return HRESULT.S_OK;
-            }
-
-            unsafe HRESULT IDocHostUIHandler.TranslateAccelerator(User32.MSG* lpMsg, Guid* pguidCmdGroup, uint nCmdID)
-            {
-                if (lpMsg is null || pguidCmdGroup is null)
-                {
-                    return HRESULT.E_POINTER;
-                }
-
-                // Returning S_FALSE will allow the native control to do default processing,
-                // i.e., execute the shortcut key. Returning S_OK will cancel the shortcut key.
-                WebBrowser wb = (WebBrowser)Host;
-                if (!wb.WebBrowserShortcutsEnabled)
-                {
-                    int keyCode = (int)lpMsg->wParam | (int)Control.ModifierKeys;
-                    if (lpMsg->message != User32.WM.CHAR && Enum.IsDefined(typeof(Shortcut), (Shortcut)keyCode))
-                    {
-                        return HRESULT.S_OK;
-                    }
-                }
-
-                return HRESULT.S_FALSE;
-            }
-
-            HRESULT IDocHostUIHandler.TranslateUrl(uint dwTranslate, string strUrlIn, out string pstrUrlOut)
-            {
-                // Set to null no matter what we return, to prevent the marshaller
-                // from having issues if the pointer points to random stuff.
-                pstrUrlOut = null;
-                return HRESULT.S_FALSE;
-            }
-
-            HRESULT IDocHostUIHandler.FilterDataObject(IComDataObject pDO, out IComDataObject ppDORet)
-            {
-                // Set to null no matter what we return, to prevent the marshaller
-                // from having issues if the pointer points to random stuff.
-                ppDORet = null;
-                return HRESULT.S_FALSE;
-            }
-
-            //
-            // Internal methods
-            //
-            internal override void OnPropertyChanged(Ole32.DispatchID dispid)
-            {
-                if (dispid != Ole32.DispatchID.READYSTATE)
-                {
-                    base.OnPropertyChanged(dispid);
-                }
-            }
-        }
-
-        [ClassInterface(ClassInterfaceType.None)]
-        private class WebBrowserEvent : StandardOleMarshalObject, SHDocVw.DWebBrowserEvents2
-        {
-            private readonly WebBrowser _parent;
-            private bool _haveNavigated;
-
-            public WebBrowserEvent(WebBrowser parent)
-            {
-                _parent = parent;
-            }
-
-            public bool AllowNavigation { get; set; }
-
-            public void CommandStateChange(SHDocVw.CSC command, bool enable)
-            {
-                if (command == SHDocVw.CSC.NAVIGATEBACK)
-                {
-                    _parent.CanGoBackInternal = enable;
-                }
-                else if (command == SHDocVw.CSC.NAVIGATEFORWARD)
-                {
-                    _parent.CanGoForwardInternal = enable;
-                }
-            }
-
-            public void BeforeNavigate2(object pDisp, ref object urlObject, ref object flags, ref object targetFrameName, ref object postData, ref object headers, ref bool cancel)
-            {
-                Debug.Assert(_parent != null, "Parent should have been set");
-                //Note: we want to allow navigation if we haven't already navigated.
-                if (AllowNavigation || !_haveNavigated)
-                {
-                    Debug.Assert(urlObject is null || urlObject is string, "invalid url type");
-                    Debug.Assert(targetFrameName is null || targetFrameName is string, "invalid targetFrameName type");
-                    Debug.Assert(headers is null || headers is string, "invalid headers type");
-                    //
-                    // If during running interop code, the variant.bstr value gets set
-                    // to -1 on return back to native code, if the original value was null, we
-                    // have to set targetFrameName and headers to "".
-                    if (targetFrameName is null)
-                    {
-                        targetFrameName = string.Empty;
-                    }
-                    if (headers is null)
-                    {
-                        headers = string.Empty;
-                    }
-
-                    string urlString = urlObject is null ? "" : (string)urlObject;
-                    WebBrowserNavigatingEventArgs e = new WebBrowserNavigatingEventArgs(
-                        new Uri(urlString), targetFrameName is null ? "" : (string)targetFrameName);
-                    _parent.OnNavigating(e);
-                    cancel = e.Cancel;
-                }
-                else
-                {
-                    cancel = true;
-                }
-            }
-
-            public void DocumentComplete(object pDisp, ref object urlObject)
-            {
-                Debug.Assert(urlObject is null || urlObject is string, "invalid url");
-                _haveNavigated = true;
-                if (_parent.documentStreamToSetOnLoad != null && (string)urlObject == "about:blank")
-                {
-                    HtmlDocument htmlDocument = _parent.Document;
-                    if (htmlDocument != null)
-                    {
-                        Ole32.IPersistStreamInit psi = htmlDocument.DomDocument as Ole32.IPersistStreamInit;
-                        Debug.Assert(psi != null, "The Document does not implement IPersistStreamInit");
-                        Ole32.IStream iStream = (Ole32.IStream)new Ole32.GPStream(
-                                                    _parent.documentStreamToSetOnLoad);
-                        psi.Load(iStream);
-                        htmlDocument.Encoding = "unicode";
-                    }
-                    _parent.documentStreamToSetOnLoad = null;
-                }
-                else
-                {
-                    string urlString = urlObject is null ? "" : urlObject.ToString();
-                    WebBrowserDocumentCompletedEventArgs e = new WebBrowserDocumentCompletedEventArgs(
-                            new Uri(urlString));
-                    _parent.OnDocumentCompleted(e);
-                }
-            }
-
-            public void TitleChange(string text)
-            {
-                _parent.OnDocumentTitleChanged(EventArgs.Empty);
-            }
-
-            public void SetSecureLockIcon(int secureLockIcon)
-            {
-                _parent.encryptionLevel = (WebBrowserEncryptionLevel)secureLockIcon;
-                _parent.OnEncryptionLevelChanged(EventArgs.Empty);
-            }
-
-            public void NavigateComplete2(object pDisp, ref object urlObject)
-            {
-                Debug.Assert(urlObject is null || urlObject is string, "invalid url type");
-                string urlString = urlObject is null ? "" : (string)urlObject;
-                WebBrowserNavigatedEventArgs e = new WebBrowserNavigatedEventArgs(
-                        new Uri(urlString));
-                _parent.OnNavigated(e);
-            }
-
-            public void NewWindow2(ref object ppDisp, ref bool cancel)
-            {
-                CancelEventArgs e = new CancelEventArgs();
-                _parent.OnNewWindow(e);
-                cancel = e.Cancel;
-            }
-
-            public void ProgressChange(int progress, int progressMax)
-            {
-                WebBrowserProgressChangedEventArgs e = new WebBrowserProgressChangedEventArgs(progress, progressMax);
-                _parent.OnProgressChanged(e);
-            }
-
-            public void StatusTextChange(string text)
-            {
-                _parent.statusText = text ?? string.Empty;
-                _parent.OnStatusTextChanged(EventArgs.Empty);
-            }
-
-            public void DownloadBegin() => _parent.OnFileDownload(EventArgs.Empty);
-
-            public void FileDownload(ref bool cancel)
-            {
-            }
-
-            public void PrivacyImpactedStateChange(bool bImpacted)
-            {
-            }
-
-            public void UpdatePageStatus(object pDisp, ref object nPage, ref object fDone)
-            {
-            }
-
-            public void PrintTemplateTeardown(object pDisp)
-            {
-            }
-
-            public void PrintTemplateInstantiation(object pDisp)
-            {
-            }
-
-            public void NavigateError(object pDisp, ref object url, ref object frame, ref object statusCode, ref bool cancel)
-            {
-            }
-
-            public void ClientToHostWindow(ref long cX, ref long cY)
-            {
-            }
-
-            public void WindowClosing(bool isChildWindow, ref bool cancel)
-            {
-            }
-
-            public void WindowSetHeight(int height)
-            {
-            }
-
-            public void WindowSetWidth(int width)
-            {
-            }
-
-            public void WindowSetTop(int top)
-            {
-            }
-
-            public void WindowSetLeft(int left)
-            {
-            }
-
-            public void WindowSetResizable(bool resizable)
-            {
-            }
-
-            public void OnTheaterMode(bool theaterMode)
-            {
-            }
-
-            public void OnFullScreen(bool fullScreen)
-            {
-            }
-
-            public void OnStatusBar(bool statusBar)
-            {
-            }
-
-            public void OnMenuBar(bool menuBar)
-            {
-            }
-
-            public void OnToolBar(bool toolBar)
-            {
-            }
-
-            public void OnVisible(bool visible)
-            {
-            }
-
-            public void OnQuit()
-            {
-            }
-
-            public void PropertyChange(string szProperty)
-            {
-            }
-
-            public void DownloadComplete()
-            {
             }
         }
     }
